Fix connect system test echo bug without sleeps (#424)
* Remove quotes around the string to echo in connect tests

This prevents the system test from failing if it starts sending input
too quickly and only part of the echo string is actually sent to the
terminal. In these cases the system tests would never actually execute
a command because there would be an unterminated double quote in the
terminal that system test would never terminate. Therefore the command
would never get executed and the test would time out. Instead now if
we send input too quickly the first command may not be a properly
formatted echo and the test expecatation on stdout will not
pass. However we will just try sending the command again. This removes
the need for artificial sleep statements we put in to try and
otherwise prevent us from sending input too quickly to stdin.